修复新用户创建会话重置的bug，添加开源许可

+++ server/src/main/java/com/mola/molachat/controller/ChatterController.java
@@ -164,6 +164,9 @@ public class ChatterController {
      */
     @DeleteMapping
     public ServerResponse deletePreChatter(@RequestParam("preId") String preId) {
+        if (preId.length() == 0) {
+            return ServerResponse.createBySuccess();
+        }
         ChatterDTO chatterDTO = new ChatterDTO();
         chatterDTO.setId(preId);
         chatterService.remove(chatterDTO);
